migration_planning/requirements.txt
pandas
numpy
networkx
python-louvain
python-igraph
leidenalg
matplotlib
seaborn
graphviz
plotly
scikit-learn
tqdm
pyyaml
